whatsapp-temp/requirements.txt
appdirs==1.4.4
attrs==20.3.0
bitrix24-rest==1.1.1
cached-property==1.5.2
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
defusedxml==0.6.0
Deprecated==1.2.10
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.6.1
MarkupSafe==1.1.1
pycep-correios==4.0.4
PyJWT==1.7.1
pytz==2020.4
requests==2.25.0
requests-file==1.5.1
requests-toolbelt==0.9.1
six==1.15.0
twilio==6.47.0
urllib3==1.26.1
Werkzeug==1.0.1
wrapt==1.12.1
zeep==4.0.0
